pyproject.toml
[project]
name = "youtube-shorts-maker"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-adk>=1.26.0",
    "google-genai>=1.65.0",
    "litellm>=1.82.0",
    "openai>=2.24.0",
    "pydantic>=2.12.5",
]
